fix(rtems5-leon3-nexysa7): build with the strict C11 flags
Under -std=c11 newlib hides the BSD declarations, and rtems/confdefs.h needs
struct bintime, so the platform Makefile now defines _DEFAULT_SOURCE, as the
POSIX one already does. The build also reported an unused variable in the
message queue receive, a comparison between the unsigned microseconds of a
TimeVal and the signed ones of a struct timeval in the timer back end, and the
unused argument of the Init task.

+++ os/rtems5/src/periodic_timer.c
@@ -45,7 +45,7 @@ static rtems_interval get_sleep_time(const TimeVal * const next_time) {
         interval.tv_sec = 0;
         interval.tv_usec = 0;
 
-        if (next_time->tv_usec >= current_time.tv_usec) {
+        if (next_time->tv_usec >= (uint32_t)current_time.tv_usec) {
 
             interval.tv_sec = next_time->tv_sec - current_time.tv_sec;
             interval.tv_usec = next_time->tv_usec - current_time.tv_usec;
@@ -72,7 +72,7 @@ static rtems_interval get_sleep_time(const TimeVal * const next_time) {
 
     } else if (next_time->tv_sec == current_time.tv_sec) {
 
-        if (next_time->tv_usec > current_time.tv_usec) {
+        if (next_time->tv_usec > (uint32_t)current_time.tv_usec) {
 
             TimeVal interval;
